+++ codegen/src/worktable/model/column.rs
@@ -1,13 +1,14 @@
 use std::collections::HashMap;
 
 use proc_macro2::{Ident, TokenStream};
+use quote::quote;
 use syn::spanned::Spanned;
 use crate::worktable::model::GeneratorType;
 use crate::worktable::model::index::Index;
 
 #[derive(Debug, Clone)]
 pub struct Columns {
-    pub columns_map: HashMap<Ident, Ident>,
+    pub columns_map: HashMap<Ident, TokenStream>,
     pub indexes: HashMap<Ident, Index>,
     pub primary_keys: Vec<Ident>,
     pub generator_type: GeneratorType
@@ -19,6 +20,7 @@ pub struct Row {
     pub type_: Ident,
     pub is_primary_key: bool,
     pub gen_type: GeneratorType,
+    pub optional: bool
 }
 
 impl Columns {
@@ -28,7 +30,13 @@ impl Columns {
         let mut gen_type = None;
 
         for row in rows {
-            columns_map.insert(row.name.clone(), row.type_.clone());
+            let type_ = &row.type_;
+            let type_  = if row.optional {
+                quote! { core::option::Option<#type_> }
+            } else {
+                quote! { #type_ }
+            };
+            columns_map.insert(row.name.clone(), type_);
 
             if row.is_primary_key {
                 if let Some(t) = gen_type {