+++ codegen/src/worktable/model/primary_key.rs
@@ -1,10 +1,10 @@
 use std::collections::HashMap;
-use proc_macro2::Ident;
+use proc_macro2::{Ident, TokenStream};
 
 #[derive(Debug, Clone)]
 pub struct PrimaryKey {
     pub ident: Ident,
-    pub vals: HashMap<Ident, Ident>,
+    pub vals: HashMap<Ident, TokenStream>,
 }
 
 #[derive(Debug, Clone, Copy, PartialEq)]